Omaha #8723 Simplify keystore/truststore handling
Cherry-picked from dev. Here is the original commit message:

---

Total redo of keystore/truststore management for Ignite and Qpid SSL
connections. These stores are now ephemeral (sort of), taking advantage of the
fact that they are generated from the .crt/.key files, do not contain
any additional info that needs to be preserved, and are only used by the
program that created them. I would keep them in RAM if I could but I do not
think that is supported by Qpid or Ignite.

Changes:
- The keystore and truststore files are now written to the default temp dir
  using partially-random filenames (instead of to the same dir as the crt/key
  with fixed filenames).
- Keystore and truststore are deleted on JVM exit. Thus new cert/key files will
  always be picked up on JVM restart and we do not rely on checking file
  timestamps
- The keystore and truststore passwords are now generated randomly when the
  files are created, and the passwords are kept in RAM.
- passwords.properties is no longer needed so I removed it and all code that
  manages it.
- JMS and Ignite SSL configs now share almost all code. The only substantive
  difference between the two is what locations they look in for crt/key files.
- Replaced JKS stores with PKCS12 stores. (JKS is considered a legacy format)

Change-Id: Ie0678c66c269f21b398c15ba5cb9394b40891e23