Omaha #8946 FSI: Update map and height scale menus to display correctly for new combo editor
Previously there was one scale menu and property, with map items or height menus contributed
to it depending on the type of editor active. This change splits the menu into separate map
and height menus/properties. Both menus are always displayed but are enabled/disabled depending
on if any of the panes in the active editor are of the appropriate type.

Change-Id: I54634c0a1f35ce0503413776c3c15a89b91e7533