Omaha #9020 SystemUtil.getLocalAddress skip RHEL 8 dummy interface
EDEX's ClusteredContexts are all showing as being activated on "ev".  The cause
of this problem is a network configuration change in RHEL 8, specifically that
the "ev" hostname is now associated with a new kind of "dummy" network
interface rather than a "loopback" interface as it was in RHEL 7. This is not
accounted for in the code of SystemUtil.getLocalAddress, so that method may
return the IP address from the dummy interface.

This commit fixes the problem by adding code to SystemUtil.getLocalAddress to
identify and skip the dummy interface by checking for a /32 subnet mask.

Change-Id: I1dcdd5a85cb3f170c0f5053fa50dcca0d5d56cf7